Renamed stack-navigator-screen-names
